research/requirements.txt
numpy
pandas
tqdm
pyyaml
matplotlib
scikit-image
opencv-python
albumentations
torch
torchvision